Make testAlphaProjection GDAL-specific
