+++ test/source_geo_base.py
@@ -558,23 +558,3 @@ class _GDALBaseSourceTest(_BaseGeoTests):
             imagePath, projection='EPSG:3857',
             style={'bands': [{'band': 1, 'max': 100, 'min': 5, 'nodata': 0}]})
         assert source.getThumbnail()[0]
-
-    def testAlphaProjection(self):
-        testDir = os.path.dirname(os.path.realpath(__file__))
-        imagePath = os.path.join(testDir, 'test_files', 'rgba_geotiff.tiff')
-        source = self.open(
-            imagePath, projection='EPSG:3857')
-        base = source.getThumbnail(encoding='PNG')[0]
-        basenp = source.getThumbnail(format='numpy')[0]
-        assert numpy.count_nonzero(basenp[:, :, 3] == 255) > 30000
-        source = self.open(
-            imagePath, projection='EPSG:3857',
-            style={'bands': [
-                {'band': 1, 'palette': 'R'},
-                {'band': 2, 'palette': 'G'},
-                {'band': 3, 'palette': 'B'}]})
-        assert source.getThumbnail(encoding='PNG')[0] == base
-        assert not (source.getThumbnail(format='numpy')[0] - basenp).any()
-        source = self.open(
-            imagePath)
-        assert numpy.count_nonzero(source.getThumbnail(format='numpy')[0][:, :, 3] == 255) > 30000
